Adds endpoints user authentication for right user profile

+++ src/main/java/edu/utn/UEEDServer/controller/BackofficeController.java
@@ -2,12 +2,18 @@ package edu.utn.UEEDServer.controller;
 
 import edu.utn.UEEDServer.model.*;
 import edu.utn.UEEDServer.model.dto.ConsumersDTO;
+import static edu.utn.UEEDServer.utils.Constants.*;
+
+import edu.utn.UEEDServer.model.dto.UserDTO;
 import edu.utn.UEEDServer.service.*;
 import edu.utn.UEEDServer.utils.EntityURLBuilder;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.format.annotation.DateTimeFormat;
 import org.springframework.http.HttpStatus;
+import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
+
 
 import java.util.Date;
 import java.util.List;
@@ -41,17 +47,21 @@ public class BackofficeController {
 /* RATES ENDPOINTS */
 
     @GetMapping("/rate")
-    public List<Rate> getAllRate() {
+    public List<Rate> getAllRate(Authentication auth) {
+        if(!((UserDTO) auth.getPrincipal()).getEmployee())
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN,"Access forbidden for your profile.");
         return rateService.getAll();
     }
 
     @GetMapping("/rate/{id}")
-    public Rate getByIdRate(@PathVariable Integer id) {
+    public Rate getByIdRate(Authentication auth, @PathVariable Integer id) {
+        if(!((UserDTO) auth.getPrincipal()).getEmployee())
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN,"Access forbidden for your profile.");
         return rateService.getById(id);
     }
 
     @PostMapping("/rate")
-    public PostResponse addRate(@RequestBody Rate rate) {
+    public PostResponse addRate(Authentication auth, @RequestBody Rate rate) {
         Rate added = rateService.add(rate);
         return PostResponse.builder().
                 status(HttpStatus.CREATED).
@@ -60,7 +70,9 @@ public class BackofficeController {
     }
 
     @PutMapping("/rate")
-    public PostResponse updateRate(@RequestBody Rate rate) {
+    public PostResponse updateRate(Authentication auth, @RequestBody Rate rate) {
+        if(!((UserDTO) auth.getPrincipal()).getEmployee())
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN,"Access forbidden for your profile.");
 
         Rate saved = rateService.update(rate);
 
@@ -72,7 +84,9 @@ public class BackofficeController {
 
     // Shall we implement DELETE method??
     @DeleteMapping("/rate/{id}")
-    public void deleteRate(@PathVariable Integer id) {
+    public void deleteRate(Authentication auth, @PathVariable Integer id) {
+        if(!((UserDTO) auth.getPrincipal()).getEmployee())
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN,"Access forbidden for your profile.");
         rateService.delete(id);
     }
 
@@ -80,17 +94,21 @@ public class BackofficeController {
 /* ADDRESSES ENDPOINTS */
 
     @GetMapping("/address")
-    public List<Address> getAllAddress() {
+    public List<Address> getAllAddress(Authentication auth) {
         return addressService.getAll();
     }
 
     @GetMapping("/address/{id}")
-    public Address getByIdAddress(@PathVariable Integer id) {
+    public Address getByIdAddress(Authentication auth, @PathVariable Integer id) {
+        if(!((UserDTO) auth.getPrincipal()).getEmployee())
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN,"Access forbidden for your profile.");
         return addressService.getById(id);
     }
 
     @PostMapping("/client/{clientId}/address")
-    public PostResponse addAddress(@PathVariable Integer clientId, @RequestBody Address address) {
+    public PostResponse addAddress(Authentication auth, @PathVariable Integer clientId, @RequestBody Address address) {
+        if(!((UserDTO) auth.getPrincipal()).getEmployee())
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN,"Access forbidden for your profile.");
         Address newAddress = addressService.add(clientId, address);
 
         return PostResponse.builder().
@@ -100,7 +118,9 @@ public class BackofficeController {
     }
 
     @PutMapping("/client/{clientId}/address")
-    public PostResponse updateAddress(@PathVariable Integer clientId, @RequestBody Address address) {
+    public PostResponse updateAddress(Authentication auth, @PathVariable Integer clientId, @RequestBody Address address) {
+        if(!((UserDTO) auth.getPrincipal()).getEmployee())
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN,"Access forbidden for your profile.");
         Client updated = addressService.update(clientId, address);
 
         return PostResponse.builder().
@@ -110,7 +130,9 @@ public class BackofficeController {
     }
 
     @DeleteMapping("/address/{addressId}")
-    public void deleteAddress(@PathVariable Integer addressId) {
+    public void deleteAddress(Authentication auth, @PathVariable Integer addressId) {
+        if(!((UserDTO) auth.getPrincipal()).getEmployee())
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN,"Access forbidden for your profile.");
         addressService.delete(addressId);
     }
 
@@ -118,17 +140,23 @@ public class BackofficeController {
 /* METER ENDPOINTS */
 
     @GetMapping
-    public List<Meter> getAllMeter() {
+    public List<Meter> getAllMeter(Authentication auth) {
+        if(!((UserDTO) auth.getPrincipal()).getEmployee())
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN,"Access forbidden for your profile.");
         return this.meterService.getAll();
     }
 
     @GetMapping("/meter/{serialNumber}")
-    public Meter getByIdMeter(@PathVariable String serialNumber) {
+    public Meter getByIdMeter(Authentication auth, @PathVariable String serialNumber) {
+        if(!((UserDTO) auth.getPrincipal()).getEmployee())
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN,"Access forbidden for your profile.");
         return this.meterService.getById(serialNumber);
     }
 
     @PostMapping("/meter")
-    public PostResponse addMeter(@RequestBody Meter newMeter) {
+    public PostResponse addMeter(Authentication auth, @RequestBody Meter newMeter) {
+        if(!((UserDTO) auth.getPrincipal()).getEmployee())
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN,"Access forbidden for your profile.");
         Meter saved = meterService.add(newMeter);
 
         return PostResponse.builder().
@@ -138,7 +166,9 @@ public class BackofficeController {
     }
 
     @PutMapping("/meter")
-    public PostResponse updateMeter(@RequestBody Meter meter) {
+    public PostResponse updateMeter(Authentication auth, @RequestBody Meter meter) {
+        if(!((UserDTO) auth.getPrincipal()).getEmployee())
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN,"Access forbidden for your profile.");
         Meter m = meterService.update(meter);
 
         return PostResponse.builder().
@@ -148,7 +178,9 @@ public class BackofficeController {
     }
 
     @DeleteMapping("/meter/{serialNumber}")
-    public void deleteMeter(@PathVariable String serialNumber) {
+    public void deleteMeter(Authentication auth, @PathVariable String serialNumber) {
+        if(!((UserDTO) auth.getPrincipal()).getEmployee())
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN,"Access forbidden for your profile.");
         this.meterService.delete(serialNumber);
     }
 
@@ -156,28 +188,38 @@ public class BackofficeController {
 /* BILL ENDPOINTS */
 
     @GetMapping("/client/{clientId}/bills") // VER QUERY DSL PARA LOS DIFERENTES FILTERS
-    public List<Bill> filterByDate(@PathVariable Integer clientId,
+    public List<Bill> filterByDate(Authentication auth,
+                                   @PathVariable Integer clientId,
                                    @RequestParam @DateTimeFormat(pattern="yyyy-MM") Date from,
                                    @RequestParam @DateTimeFormat(pattern="yyyy-MM") Date to){
+        if(!((UserDTO) auth.getPrincipal()).getEmployee())
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN,"Access forbidden for your profile.");
         return this.billService.filterByClientAndDate(clientId,from,to);
     }
 
     @GetMapping("/client/{clientId}/bills/unpaid")
-    public List<Bill>getUnpaidBillClient(@PathVariable Integer clientId){
+    public List<Bill>getUnpaidBillClient(Authentication auth, @PathVariable Integer clientId){
+        if(!((UserDTO) auth.getPrincipal()).getEmployee())
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN,"Access forbidden for your profile.");
         return billService.getClientUnpaid(clientId);
     }
 
     @GetMapping("/address/{addressId}/bills/unpaid")
-    public List<Bill>getUnpaidBillAddress(@PathVariable Integer addressId){
+    public List<Bill>getUnpaidBillAddress(Authentication auth, @PathVariable Integer addressId){
+        if(!((UserDTO) auth.getPrincipal()).getEmployee())
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN,"Access forbidden for your profile.");
             return billService.getAddressUnpaid(addressId);
     }
 
 /* READING ENDPOINTS */
 
     @GetMapping("/address/{addressId}/readings")
-    public List<Reading> getAddressReadings(@PathVariable Integer addressId,
+    public List<Reading> getAddressReadings(Authentication auth,
+                                            @PathVariable Integer addressId,
                                             @RequestParam @DateTimeFormat(pattern="yyyy-MM-dd") Date from,
                                             @RequestParam @DateTimeFormat(pattern="yyyy-MM-dd") Date to){
+        if(!((UserDTO) auth.getPrincipal()).getEmployee())
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN,"Access forbidden for your profile.");
         return this.readingService.getAddressReadingsByDate(addressId,from,to);
     }
 
@@ -185,9 +227,11 @@ public class BackofficeController {
 /* CLIENT ENDPOINTS */
 
     @GetMapping("/client/topconsumers")
-    public List<ConsumersDTO> getTopConsumers( // Hacer un proyection
+    public List<ConsumersDTO> getTopConsumers( Authentication auth,
                                                @RequestParam("from") @DateTimeFormat(pattern = "yyyy-MM-dd") Date from,
                                                @RequestParam("to") @DateTimeFormat(pattern = "yyyy-MM-dd") Date to){
+        if(!((UserDTO) auth.getPrincipal()).getEmployee())
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN,"Access forbidden for your profile.");
         return this.readingService.getTopConsumers(from,to);
     }
 }
